Add background to the blocklist
